Moved validator classes.
