+++ docroot/modules/custom/civictheme_migrate/src/Validator/MigrationSchemaManagerInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\civictheme_migrate;
+namespace Drupal\civictheme_migrate\Validator;
 
 /**
  * Manage migration schemas.
@@ -10,7 +10,7 @@ interface MigrationSchemaManagerInterface {
   /**
    * Get all migration schemas.
    *
-   * @return \Drupal\civictheme_migrate\MigrationSchema[]
+   * @return \Drupal\civictheme_migrate\Validator\MigrationSchema[]
    *   Array of migration schemas.
    */
   public function getSchemas(): array;
@@ -21,7 +21,7 @@ interface MigrationSchemaManagerInterface {
    * @param string $id
    *   The migration schema ID.
    *
-   * @return \Drupal\civictheme_migrate\MigrationSchema|null
+   * @return \Drupal\civictheme_migrate\Validator\MigrationSchema|null
    *   Migration schema of found or NULL otherwise.
    */
   public function getSchema(string $id): MigrationSchema|null;
